requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
supabase==2.0.3
python-dotenv==1.0.0
